Cargo

+++ src/main/java/com/guardias/backend/dto/AdicionalDto.java
@@ -5,23 +5,23 @@ import jakarta.validation.constraints.NotBlank;
 public class AdicionalDto {
 
     @NotBlank
-    private int id;
+    private long id;
     @NotBlank
     private String nombre;
 
     public AdicionalDto() {
     }
 
-    public AdicionalDto(int id, String nombre) {
+    public AdicionalDto(long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
