+++ src/main/java/com/guardias/backend/dto/CategoriaDto.java
@@ -5,23 +5,23 @@ import jakarta.validation.constraints.NotBlank;
 public class CategoriaDto {
 
     @NotBlank
-    private int id;
+    private long id;
     @NotBlank
     private String nombre;
 
     public CategoriaDto() {
     }
 
-    public CategoriaDto(@NotBlank int id, @NotBlank String nombre) {
+    public CategoriaDto(@NotBlank long id, @NotBlank String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
